requirements.txt
pytest
pytest-pep8
requests
numpy
matplotlib
Pillow
scikit-image
